Split tests according to classes
