+++ src/test/resources/application.properties
@@ -11,6 +11,8 @@ spring.jpa.hibernate.ddl-auto=create-drop
 spring.profiles.active=test
 spring.threads.virtual.enabled=true
 
+server.servlet.context-path=/api
+
 server.ssl.certificate-header-name=${HEADER_NAME:X-APP-CERTIFICATE}
 
 scheduled-tasks.enabled=false
